Fix NavGraph initialization and resolve issue with Dagger Map.
### Problem:
The App module couldn't access `home_api` interfaces because they were declared as `implementation` in `home_impl`. As a result, Dagger was unable to generate a valid `Map` for navigation, and the key in the `Map` was broken.

### Solution:
Changed the `implementation` keyword to `api` in `home_impl`'s `build.gradle`. This made the `home_api` module accessible to the App module and allowed Dagger to correctly create the `Map` with valid keys.

### Result:
NavGraph now initializes correctly, and navigation between screens works as expected.

Also adding atoms Buttons.kt